+++ tests/testthat/test-tidy.R
@@ -14,7 +14,7 @@ test_that("tidy: minimal", {
 })
 
 test_that("glance: with modelsummary", {
-    skip_if_not_installed("modelsummary", minimum_version = "0.9.2.9000")
+    skip_if_not_installed("modelsummary", minimum_version = "0.9.3")
     gl <- glance(mfx)
     expect_equal(dim(glance(mfx)), c(1, 9))
 })